shared/requirements.txt
aiohttp
pydantic
asyncpg
psycopg2-binary
python-dotenv
azure-storage-blob
azure-storage-queue
azure-identity
azure-messaging-servicebus
